Fix grid content indentation regression
In #b43b47f I had removed the overall left margin for .content a's and li's as this caused issues on the orders page where a link is next to order headings.

That did however break the indentation of grid cell items on pages like the profile and battalion ones.

This commit adds that indentation back but limits it to .content .row's as I'm not using .cols everywhere but more specific widths in places.